move seniority logic to own module
