+++ mobile_verifier/src/lib.rs
@@ -8,6 +8,7 @@ pub mod heartbeats;
 pub mod radio_threshold;
 pub mod reward_shares;
 pub mod rewarder;
+pub mod seniority;
 mod settings;
 pub mod sp_boosted_rewards_bans;
 pub mod speedtests;
