+++ mobile_verifier/tests/integrations/modeled_coverage.rs
@@ -15,14 +15,14 @@ use hextree::Cell;
 use mobile_config::boosted_hex_info::{BoostedHexInfo, BoostedHexes};
 
 use mobile_verifier::{
-    coverage::{CoverageClaimTimeCache, CoverageObject, CoverageObjectCache, Seniority},
+    coverage::{CoverageClaimTimeCache, CoverageObject, CoverageObjectCache},
     geofence::GeofenceValidator,
     heartbeats::{
-        last_location::LocationCache, Heartbeat, HeartbeatReward, KeyType, SeniorityUpdate,
-        ValidatedHeartbeat,
+        last_location::LocationCache, Heartbeat, HeartbeatReward, KeyType, ValidatedHeartbeat,
     },
     reward_shares::CoverageShares,
     rewarder::boosted_hex_eligibility::BoostedHexEligibility,
+    seniority::{Seniority, SeniorityUpdate},
     speedtests::Speedtest,
     speedtests_average::{SpeedtestAverage, SpeedtestAverages},
     GatewayResolution, GatewayResolver, IsAuthorized,
